# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ForecastHorizon": {
        "enum": [
          "day",
          "month",
          "year"
        ],
        "title": "ForecastHorizon",
        "type": "string"
      },
...
      "ForecastMapResponse": {
        "properties": {
          "ambiguous_count": {
            "title": "Ambiguous Count",
            "type": "integer"
          },
          "entity_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Version"
          },
          "graph_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graph Version"
          },
          "mapping_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mapping Version"
          },
          "matched_count": {
            "title": "Matched Count",
            "type": "integer"
          },
          "positions": {
            "items": {
              "$ref": "#/components/schemas/ForecastMapPositionResponse"
            },
            "title": "Positions",
            "type": "array"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "status": {
            "enum": [
              "ready",
              "partial",
              "unavailable"
            ],
            "title": "Status",
            "type": "string"
          },
          "synthetic": {
            "title": "Synthetic",
            "type": "boolean"
          },
          "unmatched_count": {
            "title": "Unmatched Count",
            "type": "integer"
          }
        },
        "required": [
          "run_id",
          "entity_version",
          "graph_version",
          "mapping_version",
          "synthetic",
          "status",
          "reason",
          "matched_count",
          "unmatched_count",
          "ambiguous_count",
          "positions"
        ],
        "title": "ForecastMapResponse",
        "type": "object"
      },
      "ForecastPointResponse": {
        "properties": {
          "capacity": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity"
          },
          "lower_bound": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lower Bound"
          },
          "predicted_passengers": {
            "minimum": 0.0,
            "title": "Predicted Passengers",
            "type": "number"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "upper_bound": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upper Bound"
          }
        },
        "required": [
          "timestamp",
          "predicted_passengers",
          "lower_bound",
          "upper_bound",
          "capacity"
        ],
        "title": "ForecastPointResponse",
        "type": "object"
      },
      "ForecastResponse": {
        "properties": {
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "horizon": {
            "$ref": "#/components/schemas/ForecastHorizon"
          },
          "map": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ForecastMapResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "peak_load_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peak Load Percent"
          },
          "peak_passengers": {
            "title": "Peak Passengers",
            "type": "number"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/ForecastPointResponse"
            },
            "title": "Points",
            "type": "array"
          },
          "route": {
            "$ref": "#/components/schemas/RouteResponse"
          },
          "run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ForecastRunResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "selection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ForecastSelectionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "stop_points": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/StopForecastPointResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Points"
          },
          "stops": {
            "items": {
              "$ref": "#/components/schemas/StopLoadResponse"
            },
            "title": "Stops",
            "type": "array"
          }
        },
        "required": [
          "route",
          "horizon",
          "generated_at",
          "model_version",
          "peak_passengers",
          "peak_load_percent",
          "points",
          "stops"
        ],
        "title": "ForecastResponse",
        "type": "object"
      },
      "ForecastRunResponse": {
        "properties": {
          "calendar_version": {
            "title": "Calendar Version",
            "type": "string"
          },
          "data_cutoff": {
            "format": "date-time",
            "title": "Data Cutoff",
            "type": "string"
          },
          "dataset_id": {
            "title": "Dataset Id",
            "type": "string"
          },
          "entity_version": {
            "title": "Entity Version",
            "type": "string"
          },
          "feature_version": {
            "title": "Feature Version",
            "type": "string"
          },
          "forecast_origin": {
            "format": "date-time",
            "title": "Forecast Origin",
            "type": "string"
          },
          "graph_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graph Version"
          },
          "identity_namespace": {
            "title": "Identity Namespace",
            "type": "string"
          },
          "interval_level": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interval Level"
          },
          "interval_method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interval Method"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "source_version": {
            "title": "Source Version",
            "type": "string"
          },
          "synthetic": {
            "title": "Synthetic",
            "type": "boolean"
          },
          "target": {
            "title": "Target",
            "type": "string"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "dataset_id",
          "source_version",
          "feature_version",
          "entity_version",
          "calendar_version",
          "graph_version",
          "target",
          "unit",
          "synthetic",
          "forecast_origin",
          "data_cutoff",
          "interval_level",
          "interval_method",
          "identity_namespace"
        ],
        "title": "ForecastRunResponse",
        "type": "object"
      },
      "ForecastSelectionResponse": {
        "properties": {
          "aggregation_key": {
            "title": "Aggregation Key",
            "type": "string"
          },
          "direction_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction Id"
          },
          "end": {
            "format": "date-time",
            "title": "End",
            "type": "string"
          },
          "interval_aggregation": {
            "title": "Interval Aggregation",
            "type": "string"
          },
          "start": {
            "format": "date-time",
            "title": "Start",
            "type": "string"
          },
          "stop_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Id"
          }
        },
        "required": [
          "start",
          "end",
          "stop_id",
          "direction_id",
          "aggregation_key",
          "interval_aggregation"
        ],
        "title": "ForecastSelectionResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlanningRequest": {
        "additionalProperties": false,
        "properties": {
          "direction": {
            "anyOf": [
              {
                "maxLength": 30,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction"
          },
          "factors": {
            "$ref": "#/components/schemas/Factors"
          },
          "forecast_mode": {
            "default": "approved",
            "enum": [
              "approved",
              "external_experiment",
              "stop_model"
            ],
            "title": "Forecast Mode",
            "type": "string"
          },
          "horizon": {
            "default": "day",
            "enum": [
              "day",
              "month",
              "year"
            ],
            "title": "Horizon",
            "type": "string"
          },
          "route": {
            "maxLength": 10,
            "minLength": 1,
            "title": "Route",
            "type": "string"
          },
          "source_enabled": {
            "$ref": "#/components/schemas/SourceFlags"
          },
          "start_date": {
            "format": "date",
            "title": "Start Date",
            "type": "string"
          },
          "stop_id": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Id"
          }
        },
        "required": [
          "route",
          "start_date"
        ],
        "title": "PlanningRequest",
        "type": "object"
      },
      "PlanningResponse": {
        "additionalProperties": false,
        "properties": {
          "direction": {
            "anyOf": [
              {
                "maxLength": 30,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction"
          },
          "experimental_evaluation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experimental Evaluation"
          },
          "factors": {
            "$ref": "#/components/schemas/Factors"
          },
          "forecast_mode": {
            "default": "approved",
            "enum": [
              "approved",
              "external_experiment",
              "stop_model"
            ],
            "title": "Forecast Mode",
            "type": "string"
          },
          "generated_at": {
            "title": "Generated At",
            "type": "string"
          },
          "horizon": {
            "default": "day",
            "enum": [
              "day",
              "month",
              "year"
            ],
            "title": "Horizon",
            "type": "string"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/PlanningPoint"
            },
            "title": "Points",
            "type": "array"
          },
          "provenance": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Provenance",
            "type": "object"
          },
          "qualitative": {
            "title": "Qualitative",
            "type": "boolean"
          },
          "route": {
            "maxLength": 10,
            "minLength": 1,
            "title": "Route",
            "type": "string"
          },
          "routes": {
            "items": {
              "type": "string"
            },
            "title": "Routes",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "source_enabled": {
            "$ref": "#/components/schemas/SourceFlags"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/PlanningSource"
            },
            "title": "Sources",
            "type": "array"
          },
          "start_date": {
            "format": "date",
            "title": "Start Date",
            "type": "string"
          },
          "stop_id": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Id"
          },
          "stops": {
            "items": {
              "$ref": "#/components/schemas/PlanningStop"
            },
            "title": "Stops",
            "type": "array"
          },
          "timezone": {
            "const": "Europe/Moscow",
            "title": "Timezone",
            "type": "string"
          },
          "unit": {
            "const": "event_count",
            "title": "Unit",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "route",
          "start_date",
          "experimental_evaluation",
          "run_id",
          "model_version",
          "generated_at",
          "unit",
          "timezone",
          "qualitative",
          "warnings",
          "sources",
          "points",
          "stops",
          "routes",
          "provenance"
        ],
        "title": "PlanningResponse",
        "type": "object"
      },
...
      "RouteResponse": {
        "properties": {
          "color": {
            "title": "Color",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "number": {
            "title": "Number",
            "type": "string"
          }
        },
        "required": [
          "id",
          "number",
          "name",
          "color"
        ],
        "title": "RouteResponse",
        "type": "object"
      },
      "RouteStopResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "latitude",
          "longitude",
          "sequence"
        ],
        "title": "RouteStopResponse",
        "type": "object"
      },
      "ScenarioRequest": {
        "properties": {
          "additional_vehicles": {
            "default": 0,
            "maximum": 20.0,
            "minimum": -5.0,
            "title": "Additional Vehicles",
            "type": "integer"
          },
          "demand_change_percent": {
            "default": 0,
            "maximum": 300.0,
            "minimum": -90.0,
            "title": "Demand Change Percent",
            "type": "number"
          },
          "horizon": {
            "$ref": "#/components/schemas/ForecastHorizon",
            "default": "day"
          },
          "interval_change_percent": {
            "default": 0,
            "maximum": 200.0,
            "minimum": -75.0,
            "title": "Interval Change Percent",
            "type": "number"
          },
          "route_id": {
            "exclusiveMinimum": 0.0,
            "title": "Route Id",
            "type": "integer"
          }
        },
        "required": [
          "route_id"
        ],
        "title": "ScenarioRequest",
        "type": "object"
      },
      "ScenarioResponse": {
        "properties": {
          "affected_stops": {
            "items": {
              "$ref": "#/components/schemas/StopLoadResponse"
            },
            "title": "Affected Stops",
            "type": "array"
          },
          "baseline_peak_load_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Peak Load Percent"
          },
          "capacity_delta": {
            "title": "Capacity Delta",
            "type": "number"
          },
          "passenger_delta": {
            "title": "Passenger Delta",
            "type": "number"
          },
          "scenario_peak_load_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Peak Load Percent"
          },
          "solver_version": {
            "title": "Solver Version",
            "type": "string"
          }
        },
        "required": [
          "baseline_peak_load_percent",
          "scenario_peak_load_percent",
          "passenger_delta",
          "capacity_delta",
          "affected_stops",
          "solver_version"
        ],
        "title": "ScenarioResponse",
        "type": "object"
      },
...
      "StopForecastPointResponse": {
        "properties": {
          "aggregation_scope": {
            "title": "Aggregation Scope",
            "type": "string"
          },
          "bucket_end": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bucket End"
          },
          "capacity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity"
          },
          "direction_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction Id"
          },
          "lower_bound": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lower Bound"
          },
          "predicted_passengers": {
            "title": "Predicted Passengers",
            "type": "number"
          },
          "stop_id": {
            "title": "Stop Id",
            "type": "integer"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "upper_bound": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upper Bound"
          }
        },
        "required": [
          "stop_id",
          "timestamp",
          "bucket_end",
          "direction_id",
          "predicted_passengers",
          "lower_bound",
          "upper_bound",
          "capacity",
          "aggregation_scope"
        ],
        "title": "StopForecastPointResponse",
        "type": "object"
      },
      "StopLoadResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "load_percent": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Load Percent"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "predicted_passengers": {
            "minimum": 0.0,
            "title": "Predicted Passengers",
            "type": "number"
          },
          "sequence": {
            "minimum": 1.0,
            "title": "Sequence",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "latitude",
          "longitude",
          "predicted_passengers",
          "load_percent",
          "sequence"
        ],
        "title": "StopLoadResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Прогноз пассажиропотока московского наземного транспорта",
    "title": "TramFlow API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/forecasts": {
      "get": {
        "operationId": "get_forecast_api_v1_forecasts_get",
        "parameters": [
          {
            "in": "query",
            "name": "route_id",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "Route Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "horizon",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ForecastHorizon",
              "default": "day"
            }
          },
          {
            "in": "query",
            "name": "stop_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stop Id"
            }
          },
          {
            "in": "query",
            "name": "direction_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Direction Id"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Forecast",
        "tags": [
          "forecast"
        ]
      }
    },
...
    "/api/v1/planning/forecast": {
      "post": {
        "operationId": "planning_forecast_api_v1_planning_forecast_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanningRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanningResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Planning Forecast",
        "tags": [
          "planning"
        ]
      }
    },
...
    "/api/v1/routes/{route_id}/stops": {
      "get": {
        "operationId": "list_route_stops_api_v1_routes__route_id__stops_get",
        "parameters": [
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "title": "Route Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RouteStopResponse"
                  },
                  "title": "Response List Route Stops Api V1 Routes  Route Id  Stops Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Route Stops",
        "tags": [
          "forecast"
        ]
      }
    },
    "/api/v1/scenarios/evaluate": {
      "post": {
        "operationId": "evaluate_scenario_api_v1_scenarios_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evaluate Scenario",
        "tags": [
          "forecast"
        ]
      }
    },
...
  }
}